Update BK 0000904  Great Women of India Volume – 4  by Dr.S.Ramani - Grammatical corrections
Ref : https://www.ramakrishnavivekananda.info/vivekananda/volume_5/interviews/on_indian_women.htm

"I never define anything", said the Swami, smiling. "Still, it may be described as a development of faculty, not an accumulation of words, or as a training of individuals to will rightly and efficiently. So shall we bring to the need of India great fearless women — women worthy to continue the traditions of Sanghamittâ, Lilâ, Ahalyâ Bâi, and Mirâ Bâi — women fit to be mothers of heroes, because they are pure and selfless, strong with the strength that comes of touching the feet of God."