CL 0000833 Value Oriented Moral Lessons Volumes 1 – 4 by Swami Raghaveshananda
These series of books are written to inculcate faith in God in the young minds. Once such a faith is cultivated it helps the children to develop respect for parents teachers elders and our culture and rich heritage which in turn helps to bring peace and happiness to the home and society. Parents and teachers should help the little ones to memorise the prayers so that they will remain a source of strength throughout their lives.
See the individual books’ description for list of contents in each book.